Add files via upload
This is a deep dive into the problem (fragmentation) in General Systems Theory. Nobody can define a systems framework that is scale and domain invariant, which means, all these existing frameworks do not describe a fundamental organizing principle of "system-ness," that can describe both natural (Evolved from nature) and unnatural systems, where unnatural systems are defined as those "systems" that require conscious belief or participation from at least 1 conscious observer in order for that "system" order for it to persist. (Clinton) Alden's law of economics = No observers, no economy.  Gravity exists whether human believe it or not. Nature will never force an observer to believe in the unnatural.