Add files via upload
What is a system? The fact that there isn't a standardized definition of a system, that is both scale and domain invariant enabling recognition of both natural and unnatural system, has had profound ramifications as civilization is facing enormous competing crises. "The_Fragmentation_of_Systems_Thinking.pdf" is a deep dive into the historical background of the fragmentation of the field of General Systems Theory and how it has shaped the fragmentation of Systems Thinking in search of the unifying theory of system-ness. 

The overall purpose of this entire research and testing program is to validate, or falsify, the 7ES Framework as that unifying theory that answers that very question, "what is a system?".